refactor send actions
