+++ src/actions/index.ts
@@ -1,0 +1,2 @@
+export { default as onErrorPage } from './onErrorPage'
+export { default as onTestPass } from './onTestPass'
