--- src/actions/saveCommit.ts
@@ -1,7 +1,0 @@
-import * as git from '../services/git'
-
-async function saveCommit(): Promise<void> {
-  git.saveCommit('Save progress')
-}
-
-export default saveCommit
