+++ src/actions/tutorialConfig.ts
@@ -1,7 +1,7 @@
 import * as E from 'typings/error'
 import * as TT from 'typings/tutorial'
 import * as vscode from 'vscode'
-import { COMMANDS } from '../editor/commands'
+import { COMMANDS } from '../commands'
 import * as git from '../services/git'
 import { DISABLE_RUN_ON_SAVE } from '../environment'
