+++ src/actions/utils/loadWatchers.ts
@@ -1,6 +1,6 @@
 import * as chokidar from 'chokidar'
 import * as vscode from 'vscode'
-import { COMMANDS } from '../../editor/commands'
+import { COMMANDS } from '../../commands'
 import { WORKSPACE_ROOT } from '../../environment'
 
 // NOTE: vscode createFileWatcher doesn't seem to detect changes outside of vscode
